initialState를 따로 분리
초기상태를 변경해야 할 때 initialState.js에서만 변경하면 되도록 변경
