+++ redux/initialState.js
@@ -1,0 +1,16 @@
+export const initialState = {
+    houses: [],
+    inputs: {
+      location: "",
+      floor: null,
+      deposit: "",
+      monthlyFee: "",
+      maintenenceFee: "",
+      fungi: false,
+      waterPressure: false,
+      sunshine: false,
+      sink: false,
+      aircon: false,
+      photos: []
+    }
+  };
